Working Edition for Polygon!Found the bug
The bug is caused by compareTo() function implementation.
Need to make sure we are comparing that segment against
this segment and in order of that.maxY(), that.minY() against
this.minY() (or against this.maxY() depending on whether this and that
share the same minY(),
if they share same minY(), we use this.maxY() as
the other checking point,
else we use this.minY() for checking.

Be careful about the coordinate system
We use Graphical system so areaABC would have opposite sign
as in our normal coordinate system
Graphical systems are x-axis symmetry with out normal one(x-axis
mirroed) that's why we flipped sign for CWW and CW in our AreaABC
function!